Migrate to DynamicSupervisor
Since Elixir v. 1.6.1, if the children of a supervisor have to be added
dynamically, then DynamicSupervisor should be used. This changeset
migrates from Supervisor with strategy: :simple_one_for_one to
DynamicSupervisor with strategy: :one_for_one.

Elixir v. 1.6.0 has a problem with the DynamicSupervisor, so at least
v. 1.6.1 should be used.