Move "Guess number" function from controller to channel
When the user wants to guess the number of theirs opponent, the view
sends the request to the controller to calculate the result. However, to
improve UX, the page should not be rerendered. What we try to achieve
here is SPA bahaviour.

To achieve SPA behaviour, the view should send data through web sockets,
therefore through the channel. To achive this, the form should be inside
the React component, not in the EEX template.